Use a property instead of a field for the state. Switch from byte to enum - this gives us human readable output in the UI.

+++ PediatricSoft/PediatricSensorData.cs
@@ -102,7 +102,7 @@ namespace PediatricSoft
 
                     Parallel.ForEach(Sensors, sensor =>
                     {
-                        if (sensor.State == (byte)PediatricSoftConstants.SensorState.Failed)
+                        if (sensor.State == PediatricSoftConstants.SensorState.Failed)
                             App.Current.Dispatcher.Invoke(() => Sensors.Remove(sensor));
                     });
